Reoder code and classes
